Feat : New file for containing unauthenticated user endpoints

+++ routers/selectionFile.py
@@ -1,13 +1,15 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from typing import List
 from database import database
 
-from models.selectionFileSchema import SelectionFile, SelectionFileIn
-from tables import selection_files
+from models.selectionFileSchema import SelectionFile, SelectionFileIn, SelectionFileToDelete
+from tables import selection_files, users
 from token_dependencie import JWTBearer
-
+from passlib.hash import pbkdf2_sha256
 
 router = APIRouter()
+
+USER_NOT_FOUND_MESSAGE = "User not found! The email address or username is incorrect"
 
 @router.get("/cfile", response_model=List[SelectionFile], dependencies=[Depends(JWTBearer())])
 async def get_selection_file(cfile_id:int):
@@ -30,6 +32,7 @@ async def store_selection_file(sol_file : SelectionFileIn):
         status = sol_file.status,
         nbre_sol = sol_file.n_sol,
         satisfaction = sol_file.satisfaction,
+        features = sol_file.features,
         candidatesFile_id = sol_file.candidatesFile_id,
     )
     record_id = await database.execute(query)
@@ -39,13 +42,22 @@ async def store_selection_file(sol_file : SelectionFileIn):
     return {**row}
 
 
-@router.delete("/remove", status_code = status.HTTP_200_OK, dependencies=[Depends(JWTBearer())])
-async def delete_selection_file(id:int):
+@router.post("/remove/{id}", status_code = status.HTTP_200_OK, dependencies=[Depends(JWTBearer())])
+async def delete_selection_file(id:int, data:SelectionFileToDelete):
+
+    query = users.select().where(users.c.id  == data.user_id)
+    user = await database.fetch_one(query)
+
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=USER_NOT_FOUND_MESSAGE)
+    if not pbkdf2_sha256.verify(data.user_pwd, user.password):
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=USER_NOT_FOUND_MESSAGE)
+
     query = selection_files.delete().where(selection_files.c.id == id)
     await database.execute(query)
     return {
         "removed": True,
-        "message": "Operation completed successfully"
+        "message": "Operation deleted successfully"
     }
